refine folderStructures for components, rename files
