+++ libs/ui/src/components/atoms/input/inputStyle.ts
