+++ libs/ui/src/components/atoms/inputLabel/inputLabelStyle.ts
